♻️ refactor(User): UserSkill, QuizCategory 데이터 통일 (#156)
