+++ src/main/java/com/sparta/devquiz/domain/skill/entity/Skill.java
@@ -1,7 +1,7 @@
 package com.sparta.devquiz.domain.skill.entity;
 
 import com.sparta.devquiz.domain.user.entity.User;
-import com.sparta.devquiz.domain.user.enums.UserSkill;
+import com.sparta.devquiz.domain.skill.enums.UserSkill;
 import com.sparta.devquiz.global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
