+++ src/main/java/com/sparta/devquiz/domain/user/dto/response/SkillResponse.java
@@ -1,7 +1,7 @@
 package com.sparta.devquiz.domain.user.dto.response;
 
 import com.sparta.devquiz.domain.skill.entity.Skill;
-import com.sparta.devquiz.domain.user.enums.UserSkill;
+import com.sparta.devquiz.domain.skill.enums.UserSkill;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import lombok.Builder;
